+++ CHANGELOG.md
@@ -2,4 +2,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## 0.0.1
+## [0.1.0] 2015-11-15
+### Added
+- Support for default values
+- Support for schema.autoformly.type
+- Support for schema.regEx
+- Tests for all parsers
+
+## 0.0.1 - 2015-11-14
+
+[0.1.0]: https://github.com/wieldo/meteor-autoformly/compare/v0.0.1...v0.1.0